Add files via upload
credit.policy: 1 if the customer passes the risk assessment, 0 if not.
purpose: The purpose of the chalk (values are "credit_card", "debt_consolidation", "educational", "major_purchase", "small_business", and "all_other").
int.rate: The interest rate of the chalk as a proportion (a rate of 11% would be 0.11). Borrowers that LendingClub.com considers riskier will receive a higher interest rate.
installment: The monthly installment payment the borrower makes when the loan is financed.
log.annual.inc: The natural log of the borrower's stated annual income.
dti: The borrower's debt-to-income rate (loan divided by annual income.
fico: The FICO credit score of the borrower.
days.with.cr.line: Number of days the customer had an overdraft.
revol.bal: The balance at the end of a credit card billing period.
revol.util: The refunded share of the total credit.
inq.last.6mths: The number of inquiries lenders have made to the borrower in the last 6 months.
delinq.2yrs: The number of occurrences of a delinquency greater than 30 days within the last 2 years.
pub.rec: The number of negative records (bankruptcy, tax delinquency, convictions,...) of the borrower.